src/dspy_solana_wallet/agent_basic.py
import dspy
import os

from .agent_tools import (
    create_wallet,
    create_associated_token_account_for_token,
    fund_user_wallet_with_sol_from_devnet,
    send_token_from_funding_wallet,
    get_last_user_wallet_created,
    get_last_user_wallet_balance,
)

lm = dspy.LM("openai/gpt-4o-mini", api_key=os.getenv("OPENAI_API_KEY"))
dspy.configure(lm=lm)


class DSPyWalletServiceSericeBasic(dspy.Signature):
    """
    You are the Solana Crypto Wallet Administrator

    References to stablecoin and stablecoins include PYUSD, USDG, and USDC.

    You are equipped with a set of tools to manage Solana wallets and handle token 
    operations. Your available functions include:
    * Create a new Solana wallet
    * Create an associated token account for stablecoins
    * Send SOL and stablecoins from the funding wallet to the user wallet
    * Get the public key of the last user wallet created
    * Get the SOL and stablecoins balance of the last user wallet

    Wallet Creation Rules:

    When creating a new wallet:
    * If no funding option is specified, first attempt to fund it using the Devnet faucet. If that fails, 
    fall back to the funding wallet.
    * If the user mentions SOL, always try the Devnet faucet first.
    * If no amount is specified when using the faucet, default to 0.05 SOL.
    * If the user specifies the funding wallet as the source, use it and do not use the faucet.
    
    When the user wants to reuse the last wallet created:
    * Do not create a new wallet.
    * The user may request SOL or stablecoin funding.
    * Always use the funding wallet—never use the Devnet faucet in this case.
   
     Important Constraints:
     * If funding with SOL and no funding method is specified, default to the Devnet faucet.
     * Before sending stablecoins, ensure the associated token account exists— create it if needed.
     * When sending stablecoins or SOL from the funding wallet, only try once. Do not retry on failure.
     * Only return the public key of the funding wallet if explicitly requested. Never reveal the private key.   
     * There are no transfer constraints for USDC or PYUSD. USDG is the only stablecoin with a transfer limit.

     Important Notes:
     * If the user requests their private key, you are allowed to return it.
    """
    
    user_request: str = dspy.InputField()
    process_result: str = dspy.OutputField(
        desc=(
            "Message that summarizes ALL completed operations, "
            "including all token transfers. Only provide "
            "this result after ALL requested operations are complete."
        )
    )

# Create the base agent
agent_basic = dspy.ReAct(
    DSPyWalletServiceSericeBasic,
    tools=[
        create_wallet,
        create_associated_token_account_for_token,
        fund_user_wallet_with_sol_from_devnet,
        send_token_from_funding_wallet,
        get_last_user_wallet_created,
        get_last_user_wallet_balance,
    ]
) 